+++ src/navigation/types.ts
@@ -6,6 +6,7 @@ export type RootStackParamList = {
       title: string;
       composer: string
     };
-    List:undefined;
+    List: undefined;
+    Favourite: undefined;
   };